+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoView.kt
@@ -15,7 +15,8 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
     addView(
       playerView,
       ViewGroup.LayoutParams(
-        ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT
+        ViewGroup.LayoutParams.MATCH_PARENT,
+        ViewGroup.LayoutParams.MATCH_PARENT
       )
     )
   }